+++ src/main/java/com/tp1202510030/backend/companies/interfaces/rest/resources/company/CompanyResource.java
@@ -3,6 +3,6 @@ package com.tp1202510030.backend.companies.interfaces.rest.resources.company;
 public record CompanyResource(
         Long id,
         String companyName,
-        Integer taxIdentificationNumber
+        Long taxIdentificationNumber
 ) {
 }
