+++ src/main/java/com/tp1202510030/backend/companies/interfaces/rest/resources/company/CreateCompanyResource.java
@@ -2,6 +2,6 @@ package com.tp1202510030.backend.companies.interfaces.rest.resources.company;
 
 public record CreateCompanyResource(
         String companyName,
-        Integer taxIdentificationNumber
+        Long taxIdentificationNumber
 ) {
 }
